EntitiesMetadataQuery: Move into its own module
