+++ server/src/lib.rs
@@ -19,6 +19,7 @@ pub mod event;
 pub mod instance;
 pub mod license;
 pub mod message;
+pub mod metadata;
 pub mod navigation;
 pub mod notification;
 pub mod operation;
