+++ server/src/message.rs
@@ -6,6 +6,7 @@ use crate::alias::AliasMessage;
 use crate::database::Connection;
 use crate::event::EventMessage;
 use crate::license::LicenseMessage;
+use crate::metadata::MetadataMessage;
 use crate::navigation::NavigationMessage;
 use crate::notification::NotificationMessage;
 use crate::subject::SubjectsMessage;
@@ -28,6 +29,7 @@ pub enum Message {
     EntityMessage(EntityMessage),
     EventMessage(EventMessage),
     LicenseMessage(LicenseMessage),
+    MetadataMessage(MetadataMessage),
     NavigationMessage(NavigationMessage),
     NotificationMessage(NotificationMessage),
     PageMessage(PageMessage),
@@ -48,6 +50,7 @@ impl MessageResponder for Message {
             Message::EntityMessage(message) => message.handle(connection).await,
             Message::EventMessage(message) => message.handle(connection).await,
             Message::LicenseMessage(message) => message.handle(connection).await,
+            Message::MetadataMessage(message) => message.handle(connection).await,
             Message::NavigationMessage(message) => message.handle(connection).await,
             Message::NotificationMessage(message) => message.handle(connection).await,
             Message::PageMessage(message) => message.handle(connection).await,
