+++ server/src/metadata/mod.rs
@@ -1,0 +1,4 @@
+pub use messages::MetadataMessage;
+
+mod messages;
+mod model;
